+++ src/index.html
@@ -43,10 +43,12 @@
     <div class="main">
       <ul class="main__list">       
       </ul>
+      <ul class="library__list" data-page="library"></ul>
+      <!-- <div class="library">
+        
+      </div> -->
     </div>
-    <div class="library">
-      <ul class="library__list container " data-page="library"></ul>
-    </div>
+    
   </main>
   <div class="pagination">
     <div class="pagination__wrapper">
@@ -63,9 +65,9 @@
     </div>
   </div>
   <footer class="footer">
-    <div id="projectPreloader" class="preloader">
+    <!-- <div id="projectPreloader" class="preloader">
       <div class="loader"></div>
-    </div>
+    </div> -->
     <div class="footer__container">
       <h5 class="footer__text">
         <svg
