+++ src/js/preloader.js
@@ -1,8 +1,8 @@
-document.body.onload = function() {
-    setTimeout(function() {
-        let preloader = document.getElementById('projectPreloader');
-        if (!preloader.classList.contains('cancel')) {
-            preloader.classList.add('cancel');
-        }
-    },2700)
-};
+// document.body.onload = function() {
+//   setTimeout(function() {
+//     let preloader = document.getElementById('projectPreloader');
+//     if (!preloader.classList.contains('cancel')) {
+//         preloader.classList.add('cancel');
+//     }
+//   },2700)
+// };
