+++ src/scss/library.scss
@@ -1,65 +1,128 @@
-.library {
-  margin: 0 auto ;
-  margin-top: 20px;
-  width: 87.5%;
-  min-height: 1420px;
-  padding-right: 6.25%;
-  padding-left: 6.25%;
-  
-  &__list {
-    list-style-type: none;
-    margin: 0 auto;
-    margin-left: 10px;
+.library__list {
+  display: flex;
+  flex-direction: column;
+  list-style-type: none;
+}
+.library__item {
+  display: block;
+  margin-bottom: 30px;
+  width: 75vw;
+  height: auto;
+  a {
+    text-decoration: none;
   }
-  &__item {
-    min-width: 100%;
-    margin-bottom: 20px;
-  } 
-  &__link img {
-    width: 100%;
-    height: 93.5%;
-    border-radius: 5px;
+  &:nth-of-type(5n){
+    display: none;
+  }
+  &:nth-of-type(4){
+    margin-bottom: 40px;
+  }
+}
+.library__link img {
+  width: 75vw;
+  height: 120vw;
+  border-radius: 5px;
+  margin: 0 auto;
+} 
+
+.title{
+  padding-top: 5px;
+  display: block;
+  font-family: "Roboto Bold";
+  font-size: 3.45vh;
+  font-size: 3.45vw;
+  text-transform: uppercase;
+  color: $main-txt;
+  &:hover{
+    text-decoration: underline;
+  }
+  span{
+    padding-top: 2px;
+    display: block;
+    font-family: "Roboto Bold";
+    font-size: 3.5vh;
+    font-size: 3.5vw;
+    text-transform: none;
+    color: $highlighter;
+    &:hover{
+      text-decoration: underline;
+      color: $highlighter;
+    }
   }
 }
 
 @media (min-width: 768px) {
-  .library {
-    width: 87.5%;
-
-    &__list {
-      display: flex;
-      flex-wrap: wrap;
-      flex-direction: row;
-      justify-content: space-between;
-    }
-    &__item {
-      min-width: 45%;
-      border-radius: 5px;
-    } 
-    &__link img {
-      width: 100%;
-      height: auto;
-    }
-  }
+.library__list {
+  flex-wrap: wrap;
+  flex-direction: row;
 }
 
-@media (min-width: 1024px) {
-  .library {
-    width: 900px;
-    height: auto;
-    padding-right: 0;
-    padding-left: 0;
-    margin: 0 auto;
-    margin-top: 20px;
-    &__list {
-      justify-content: space-between;
-    }
-    &__item {
-      min-width: 30%;
-    } 
-    &__link img {
-      width: 280px;
-      height: 398px;
-    }
+.library__item {
+  width: 33.6vw;
+  height: 55vw;
+  margin-bottom: 30px;
+
+  &:nth-of-type(5n) {
+    display: block;
   }
+  &:nth-of-type(9) {
+    display: none;
+  }
+  &:nth-of-type(odd) {
+    margin-right: 8%;
+  }
+}
+.library__link img {
+  width: 105%;
+  height: 93.5%;
+}
+.title{
+  padding-top: 12px;
+  font-size: 1.6vh;
+  font-size: 1.5vw;
+  span{
+    padding-top: 3px;
+    font-size: 1.6vh;
+    font-size: 1.5vw;
+  }
+}
+}
+
+@media (min-width: 1024px){
+.library__list {
+    justify-content: space-between;
+}
+.library__item {
+  height: auto;
+  width: 274px;
+  margin-bottom: 30px;
+
+  &:nth-of-type(9) {
+    display: block;
+  }
+
+  &:nth-of-type(odd) {
+    margin-right: 0;
+  }
+  &:nth-of-type(1),
+  &:nth-of-type(2),
+  &:nth-of-type(4),
+  &:nth-of-type(5),
+  &:nth-of-type(7),
+  &:nth-of-type(8) {
+    margin-right: 30px;
+  }
+}
+.library__link img {
+  width: 274px;
+  height: 398px;
+}
+.title {
+    padding-top: 10px;
+    font-size: 14px;
+    span {
+      padding-top: 3px;
+      font-size: 14px;
+    }
+} 
 }
